#15764: moved speculative flash decode op and tests into dedicated directory and updated code owner
