+++ ttnn/cpp/ttnn/operations/experimental/speculative_execution/speculative_sdpa_decode/device/speculative_sdpa_decode_op.hpp
@@ -12,7 +12,7 @@
 #include "ttnn/tensor/tensor.hpp"
 #include "ttnn/cpp/ttnn/operations/ccl/ccl_host_types.hpp"
 
-namespace ttnn::operations::experimental::transformer {
+namespace ttnn::operations::experimental::speculative_execution {
 
 struct SpeculativeScaledDotProductAttentionDecode {
     const std::optional<float> lambda_;
@@ -49,4 +49,4 @@ struct SpeculativeScaledDotProductAttentionDecode {
         const std::vector<std::optional<const Tensor>>& optional_input_tensors) const;
 };
 
-}  // namespace ttnn::operations::experimental::transformer
+}  // namespace ttnn::operations::experimental::speculative_execution
