+++ ttnn/cpp/ttnn/operations/experimental/speculative_execution/speculative_sdpa_decode/speculative_sdpa_decode_pybind.hpp
@@ -6,7 +6,7 @@
 
 #include "pybind11/pybind_fwd.hpp"
 
-namespace ttnn::operations::experimental::transformer {
+namespace ttnn::operations::experimental::speculative_execution {
 
 void py_bind_speculative_sdpa_decode(pybind11::module& module);
-}  // namespace ttnn::operations::experimental::transformer
+}  // namespace ttnn::operations::experimental::speculative_execution
